kernels: GEMM pointer offsets in 64-bit — a matmul whose output holds more than 2^31 elements addressed memory it did not own (Mochi's VAE, CUDA 700 since 09-10; register 58)
Named in 19 minutes by --triton-sequential + CUDA_LAUNCH_BLOCKING=1 after two
compute-sanitizer runs (7 200 s, 18 000 s) measured nothing: aten.mm::1 of
mochi-1-preview's VAE, M=1 068 480 x N=2048 fp32 out — 2 188 247 040
elements. `stride_cm * offs_cm` with an int32 arange wraps past row
1 048 576 (triton-lang/triton#832: int32 index math is Triton's intended C
semantics; comfy-kitchen#172 fixed the same product on Triton 3.6.0 this
week, Triton-distributed#209 and FlagGems#6083 likewise).

- matmul_kernel, addmm_kernel, baddbmm kernel: row/column offsets (and the
  batch program id) promoted to tl.int64 before multiplying by a stride —
  A, B, C and bias pointers. No new constexpr: the autotune key and the
  certified directory are untouched.
- tests/unit/kernels/test_a_gemm_beyond_two_billion_elements.py: M=1 100 000
  x N=2048, K=64 fp16 (2 252 800 000 elements, C 4.5 GB), rows before/at/past
  the 2^31/N boundary against a host product. RED at 05:37 UTC on the kernel
  as it was (the same error 700), GREEN after (117 s incl. the sweep).
- Measured, not assumed: 50 matmul keys certified on card 2 before and after,
  same clock — best_ms after/before median 0.996 (0.980-1.020), 38/50 same
  config (two identical drafts agree on 34-39), deviations within tolerance;
  TinyLlama, Kokoro, Sana MultiLing (4 steps) byte-identical on card 3 with
  NBX_DISABLE_AUTOTUNE=1. GEMM-class kernel tests: 20 passed on card 3.

+++ src/neurobrix/kernels/ops/matmul.py
@@ -224,8 +224,11 @@ def matmul_kernel(
     offs_am = (pid_m * BLOCK_M + tl.arange(0, BLOCK_M)) % M
     offs_bn = (pid_n * BLOCK_N + tl.arange(0, BLOCK_N)) % N
     offs_k = tl.arange(0, BLOCK_K)
-    a_ptrs = a_ptr + (offs_am[:, None] * stride_am + offs_k[None, :] * stride_ak)
-    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :] * stride_bn)
+    # Pointer offsets in int64: an int32 row offset times a stride wraps past
+    # 2^31 elements (Mochi's VAE mm, M=1 068 480 x N=2048 — CUDA 700 at every
+    # attempt since 2026-09-10; test_a_gemm_beyond_two_billion_elements).
+    a_ptrs = a_ptr + (offs_am[:, None].to(tl.int64) * stride_am + offs_k[None, :] * stride_ak)
+    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :].to(tl.int64) * stride_bn)
 
     accumulator = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
     for k in range(0, tl.cdiv(K, BLOCK_K)):
@@ -251,7 +254,7 @@ def matmul_kernel(
     c = accumulator.to(c_ptr.dtype.element_ty)
     offs_cm = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
     offs_cn = pid_n * BLOCK_N + tl.arange(0, BLOCK_N)
-    c_ptrs = c_ptr + stride_cm * offs_cm[:, None] + stride_cn * offs_cn[None, :]
+    c_ptrs = c_ptr + stride_cm * offs_cm[:, None].to(tl.int64) + stride_cn * offs_cn[None, :].to(tl.int64)
     c_mask = (offs_cm[:, None] < M) & (offs_cn[None, :] < N)
     if EPILOGUE != 0:
         # Per-stage rounding emulation: `c` already carries round 1 (the
@@ -320,8 +323,11 @@ def addmm_kernel(
     offs_am = (pid_m * BLOCK_M + tl.arange(0, BLOCK_M)) % M
     offs_bn = (pid_n * BLOCK_N + tl.arange(0, BLOCK_N)) % N
     offs_k = tl.arange(0, BLOCK_K)
-    a_ptrs = a_ptr + (offs_am[:, None] * stride_am + offs_k[None, :] * stride_ak)
-    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :] * stride_bn)
+    # Pointer offsets in int64: an int32 row offset times a stride wraps past
+    # 2^31 elements (Mochi's VAE mm, M=1 068 480 x N=2048 — CUDA 700 at every
+    # attempt since 2026-09-10; test_a_gemm_beyond_two_billion_elements).
+    a_ptrs = a_ptr + (offs_am[:, None].to(tl.int64) * stride_am + offs_k[None, :] * stride_ak)
+    b_ptrs = b_ptr + (offs_k[:, None] * stride_bk + offs_bn[None, :].to(tl.int64) * stride_bn)
 
     accumulator = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
     for k in range(0, tl.cdiv(K, BLOCK_K)):
@@ -350,7 +356,7 @@ def addmm_kernel(
     accumulator = alpha * accumulator + beta * bias[None, :]
 
     offs_cm = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
-    c_ptrs = c_ptr + stride_cm * offs_cm[:, None] + stride_cn * offs_cn[None, :]
+    c_ptrs = c_ptr + stride_cm * offs_cm[:, None].to(tl.int64) + stride_cn * offs_cn[None, :].to(tl.int64)
     c_mask = (offs_cm[:, None] < M) & (offs_cn[None, :] < N)
     if EPILOGUE != 0:
         # Per-stage rounding emulation (see matmul_kernel): the unfused
